requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0